requirements.txt
python-dotenv>=1.0.0
psycopg>=3.0.0
